create util function for csv to numpy time array
